requirements.txt
django
djangorestframework
drf-yasg
whitenoise
gunicorn
Pillow
